make column reads concurrent
