+++ src/io/table/mod.rs
@@ -1,6 +1,10 @@
 pub mod index;
 pub mod key_offset;
+pub mod reader;
 pub mod table;
+pub mod writer;
 
 pub use key_offset::KeyOffset;
-pub use table::{Table, TableReader, TableWriter};
+pub use reader::TableReader;
+pub use table::Table;
+pub use writer::TableWriter;
